Move span into a separate header and add a method to read a type in LE and advance
